requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0

# LLM integration
anthropic>=0.39.0

# Environment management
python-dotenv>=1.0.0

# Markdown processing (for corpus files)
markdown>=3.5.0
